Struct aws_sdk_guardduty::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for GuardDutyAPIService.

This client allows ergonomic access to a GuardDutyAPIService-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn accept_invitation(&self) -> AcceptInvitation<C, M, R>

pub fn archive_findings(&self) -> ArchiveFindings<C, M, R>

pub fn create_detector(&self) -> CreateDetector<C, M, R>

pub fn create_filter(&self) -> CreateFilter<C, M, R>

pub fn create_ip_set(&self) -> CreateIPSet<C, M, R>

pub fn create_members(&self) -> CreateMembers<C, M, R>

pub fn create_publishing_destination(
&self
) -> CreatePublishingDestination<C, M, R>

pub fn create_sample_findings(&self) -> CreateSampleFindings<C, M, R>

pub fn create_threat_intel_set(&self) -> CreateThreatIntelSet<C, M, R>

pub fn decline_invitations(&self) -> DeclineInvitations<C, M, R>

pub fn delete_detector(&self) -> DeleteDetector<C, M, R>

pub fn delete_filter(&self) -> DeleteFilter<C, M, R>

pub fn delete_invitations(&self) -> DeleteInvitations<C, M, R>

pub fn delete_ip_set(&self) -> DeleteIPSet<C, M, R>

pub fn delete_members(&self) -> DeleteMembers<C, M, R>

pub fn delete_publishing_destination(
&self
) -> DeletePublishingDestination<C, M, R>

pub fn delete_threat_intel_set(&self) -> DeleteThreatIntelSet<C, M, R>

pub fn describe_organization_configuration(
&self
) -> DescribeOrganizationConfiguration<C, M, R>

pub fn describe_publishing_destination(
&self
) -> DescribePublishingDestination<C, M, R>

pub fn disable_organization_admin_account(
&self
) -> DisableOrganizationAdminAccount<C, M, R>

pub fn disassociate_from_master_account(
&self
) -> DisassociateFromMasterAccount<C, M, R>

pub fn disassociate_members(&self) -> DisassociateMembers<C, M, R>

pub fn enable_organization_admin_account(
&self
) -> EnableOrganizationAdminAccount<C, M, R>

pub fn get_detector(&self) -> GetDetector<C, M, R>

pub fn get_filter(&self) -> GetFilter<C, M, R>

pub fn get_findings(&self) -> GetFindings<C, M, R>

pub fn get_findings_statistics(&self) -> GetFindingsStatistics<C, M, R>

pub fn get_invitations_count(&self) -> GetInvitationsCount<C, M, R>

pub fn get_ip_set(&self) -> GetIPSet<C, M, R>

pub fn get_master_account(&self) -> GetMasterAccount<C, M, R>

pub fn get_member_detectors(&self) -> GetMemberDetectors<C, M, R>

pub fn get_members(&self) -> GetMembers<C, M, R>

pub fn get_threat_intel_set(&self) -> GetThreatIntelSet<C, M, R>

pub fn get_usage_statistics(&self) -> GetUsageStatistics<C, M, R>

pub fn invite_members(&self) -> InviteMembers<C, M, R>

pub fn list_detectors(&self) -> ListDetectors<C, M, R>

pub fn list_filters(&self) -> ListFilters<C, M, R>

pub fn list_findings(&self) -> ListFindings<C, M, R>

pub fn list_invitations(&self) -> ListInvitations<C, M, R>

pub fn list_ip_sets(&self) -> ListIPSets<C, M, R>

pub fn list_members(&self) -> ListMembers<C, M, R>

pub fn list_organization_admin_accounts(
&self
) -> ListOrganizationAdminAccounts<C, M, R>

pub fn list_publishing_destinations(
&self
) -> ListPublishingDestinations<C, M, R>

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

pub fn list_threat_intel_sets(&self) -> ListThreatIntelSets<C, M, R>

pub fn start_monitoring_members(&self) -> StartMonitoringMembers<C, M, R>

pub fn stop_monitoring_members(&self) -> StopMonitoringMembers<C, M, R>

pub fn tag_resource(&self) -> TagResource<C, M, R>

pub fn unarchive_findings(&self) -> UnarchiveFindings<C, M, R>

pub fn untag_resource(&self) -> UntagResource<C, M, R>

pub fn update_detector(&self) -> UpdateDetector<C, M, R>

pub fn update_filter(&self) -> UpdateFilter<C, M, R>

pub fn update_findings_feedback(&self) -> UpdateFindingsFeedback<C, M, R>

pub fn update_ip_set(&self) -> UpdateIPSet<C, M, R>

pub fn update_member_detectors(&self) -> UpdateMemberDetectors<C, M, R>

pub fn update_organization_configuration(
&self
) -> UpdateOrganizationConfiguration<C, M, R>

pub fn update_publishing_destination(
&self
) -> UpdatePublishingDestination<C, M, R>

pub fn update_threat_intel_set(&self) -> UpdateThreatIntelSet<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self