Module aws_sdk_guardduty::model

Modules

access_control_list

See AccessControlList

access_key_details

See AccessKeyDetails

account_detail

See AccountDetail

account_level_permissions

See AccountLevelPermissions

action

See Action

admin_account

See AdminAccount

aws_api_call_action

See AwsApiCallAction

block_public_access

See BlockPublicAccess

bucket_level_permissions

See BucketLevelPermissions

bucket_policy

See BucketPolicy

city

See City

cloud_trail_configuration_result

See CloudTrailConfigurationResult

condition

See Condition

country

See Country

data_source_configurations

See DataSourceConfigurations

data_source_configurations_result

See DataSourceConfigurationsResult

default_server_side_encryption

See DefaultServerSideEncryption

destination

See Destination

destination_properties

See DestinationProperties

dns_logs_configuration_result

See DnsLogsConfigurationResult

dns_request_action

See DnsRequestAction

domain_details

See DomainDetails

evidence

See Evidence

finding

See Finding

finding_criteria

See FindingCriteria

finding_statistics

See FindingStatistics

flow_logs_configuration_result

See FlowLogsConfigurationResult

geo_location

See GeoLocation

iam_instance_profile

See IamInstanceProfile

instance_details

See InstanceDetails

invitation

See Invitation

local_ip_details

See LocalIpDetails

local_port_details

See LocalPortDetails

master

See Master

member

See Member

member_data_source_configuration

See MemberDataSourceConfiguration

network_connection_action

See NetworkConnectionAction

network_interface

See NetworkInterface

organization

See Organization

organization_data_source_configurations

See OrganizationDataSourceConfigurations

organization_data_source_configurations_result

See OrganizationDataSourceConfigurationsResult

organization_s3_logs_configuration

See OrganizationS3LogsConfiguration

organization_s3_logs_configuration_result

See OrganizationS3LogsConfigurationResult

owner

See Owner

permission_configuration

See PermissionConfiguration

port_probe_action

See PortProbeAction

port_probe_detail

See PortProbeDetail

private_ip_address_details

See PrivateIpAddressDetails

product_code

See ProductCode

public_access

See PublicAccess

remote_ip_details

See RemoteIpDetails

remote_port_details

See RemotePortDetails

resource

See Resource

s3_bucket_detail

See S3BucketDetail

s3_logs_configuration

See S3LogsConfiguration

s3_logs_configuration_result

See S3LogsConfigurationResult

security_group

See SecurityGroup

service

See Service

sort_criteria

See SortCriteria

tag

See Tag

threat_intelligence_detail

See ThreatIntelligenceDetail

total

See Total

unprocessed_account

See UnprocessedAccount

usage_account_result

See UsageAccountResult

usage_criteria

See UsageCriteria

usage_data_source_result

See UsageDataSourceResult

usage_resource_result

See UsageResourceResult

usage_statistics

See UsageStatistics

Structs

AccessControlList

Contains information on the current access control policies for the bucket.

AccessKeyDetails

Contains information about the access keys.

AccountDetail

Contains information about the account.

AccountLevelPermissions

Contains information about the account level permissions on the S3 bucket.

Action

Contains information about actions.

AdminAccount

The account within the organization specified as the GuardDuty delegated

AwsApiCallAction

Contains information about the API action.

BlockPublicAccess

Contains information on how the bucker owner's S3 Block Public Access settings are being

BucketLevelPermissions

Contains information about the bucket level permissions for the S3 bucket.

BucketPolicy

Contains information on the current bucket policies for the S3 bucket.

City

Contains information about the city associated with the IP address.

CloudTrailConfigurationResult

Contains information on the status of CloudTrail as a data source for the detector.

Condition

Contains information about the condition.

Country

Contains information about the country where the remote IP address is located.

DataSourceConfigurations

Contains information about which data sources are enabled.

DataSourceConfigurationsResult

Contains information on the status of data sources for the detector.

DefaultServerSideEncryption

Contains information on the server side encryption method used in the S3 bucket. See

Destination

Contains information about the publishing destination, including the ID, type, and

DestinationProperties

Contains the Amazon Resource Name (ARN) of the resource to publish to, such as an S3

DnsLogsConfigurationResult

Contains information on the status of DNS logs as a data source.

DnsRequestAction

Contains information about the DNS_REQUEST action described in this finding.

DomainDetails

Contains information about the domain.

Evidence

Contains information about the reason that the finding was generated.

Finding

Contains information about the finding, which is generated when abnormal or suspicious

FindingCriteria

Contains information about the criteria used for querying findings.

FindingStatistics

Contains information about finding statistics.

FlowLogsConfigurationResult

Contains information on the status of VPC flow logs as a data source.

GeoLocation

Contains information about the location of the remote IP address.

IamInstanceProfile

Contains information about the EC2 instance profile.

InstanceDetails

Contains information about the details of an instance.

Invitation

Contains information about the invitation to become a member account.

LocalIpDetails

Contains information about the local IP address of the connection.

LocalPortDetails

Contains information about the port for the local connection.

Master

Contains information about the administrator account and invitation.

Member

Contains information about the member account.

MemberDataSourceConfiguration

Contains information on which data sources are enabled for a member account.

NetworkConnectionAction

Contains information about the NETWORK_CONNECTION action described in the finding.

NetworkInterface

Contains information about the elastic network interface of the EC2 instance.

Organization

Contains information about the ISP organization of the remote IP address.

OrganizationDataSourceConfigurations

An object that contains information on which data sources will be configured to be

OrganizationDataSourceConfigurationsResult

An object that contains information on which data sources are automatically enabled for

OrganizationS3LogsConfiguration

Describes whether S3 data event logs will be automatically enabled for new members of the

OrganizationS3LogsConfigurationResult

The current configuration of S3 data event logs as a data source for the

Owner

Contains information on the owner of the bucket.

PermissionConfiguration

Contains information about how permissions are configured for the S3 bucket.

PortProbeAction

Contains information about the PORT_PROBE action described in the finding.

PortProbeDetail

Contains information about the port probe details.

PrivateIpAddressDetails

Contains other private IP address information of the EC2 instance.

ProductCode

Contains information about the product code for the EC2 instance.

PublicAccess

Describes the public access policies that apply to the S3 bucket.

RemoteIpDetails

Contains information about the remote IP address of the connection.

RemotePortDetails

Contains information about the remote port.

Resource

Contains information about the AWS resource associated with the activity that prompted

S3BucketDetail

Contains information on the S3 bucket.

S3LogsConfiguration

Describes whether S3 data event logs will be enabled as a data source.

S3LogsConfigurationResult

Describes whether S3 data event logs will be enabled as a data source.

SecurityGroup

Contains information about the security groups associated with the EC2 instance.

Service

Contains additional information about the generated finding.

SortCriteria

Contains information about the criteria used for sorting findings.

Tag

Contains information about a tag associated with the EC2 instance.

ThreatIntelligenceDetail

An instance of a threat intelligence detail that constitutes evidence for the

Total

Contains the total usage with the corresponding currency unit for that value.

UnprocessedAccount

Contains information about the accounts that weren't processed.

UsageAccountResult

Contains information on the total of usage based on account IDs.

UsageCriteria

Contains information about the criteria used to query usage statistics.

UsageDataSourceResult

Contains information on the result of usage based on data source type.

UsageResourceResult

Contains information on the sum of usage based on an AWS resource.

UsageStatistics

Contains the result of GuardDuty usage. If a UsageStatisticType is provided the result for

Enums

AdminStatus

DataSource

DataSourceStatus

DestinationType

DetectorStatus

Feedback

FilterAction

FindingPublishingFrequency

FindingStatisticType

IpSetFormat

IpSetStatus

OrderBy

PublishingStatus

ThreatIntelSetFormat

ThreatIntelSetStatus

UsageStatisticType