Module aws_sdk_guardduty::operation

Structs

AcceptInvitation

Accepts the invitation to be monitored by a GuardDuty administrator account.

ArchiveFindings

Archives GuardDuty findings that are specified by the list of finding IDs.

CreateDetector

Creates a single Amazon GuardDuty detector. A detector is a resource that represents the

CreateFilter

Creates a filter using the specified finding criteria.

CreateIPSet

Creates a new IPSet, which is called a trusted IP list in the console user interface. An

CreateMembers

Creates member accounts of the current AWS account by specifying a list of AWS account

CreatePublishingDestination

Creates a publishing destination to export findings to. The resource to export findings to

CreateSampleFindings

Generates example findings of types specified by the list of finding types. If 'NULL' is

CreateThreatIntelSet

Creates a new ThreatIntelSet. ThreatIntelSets consist of known malicious IP addresses.

DeclineInvitations

Declines invitations sent to the current member account by AWS accounts specified by their

DeleteDetector

Deletes an Amazon GuardDuty detector that is specified by the detector ID.

DeleteFilter

Deletes the filter specified by the filter name.

DeleteIPSet

Deletes the IPSet specified by the ipSetId. IPSets are called trusted IP

DeleteInvitations

Deletes invitations sent to the current member account by AWS accounts specified by their

DeleteMembers

Deletes GuardDuty member accounts (to the current GuardDuty administrator account) specified by

DeletePublishingDestination

Deletes the publishing definition with the specified destinationId.

DeleteThreatIntelSet

Deletes the ThreatIntelSet specified by the ThreatIntelSet ID.

DescribeOrganizationConfiguration

Returns information about the account selected as the delegated administrator for

DescribePublishingDestination

Returns information about the publishing destination specified by the provided

DisableOrganizationAdminAccount

Disables an AWS account within the Organization as the GuardDuty delegated

DisassociateFromMasterAccount

Disassociates the current GuardDuty member account from its administrator account.

DisassociateMembers

Disassociates GuardDuty member accounts (to the current GuardDuty administrator account)

EnableOrganizationAdminAccount

Enables an AWS account within the organization as the GuardDuty delegated

GetDetector

Retrieves an Amazon GuardDuty detector specified by the detectorId.

GetFilter

Returns the details of the filter specified by the filter name.

GetFindings

Describes Amazon GuardDuty findings specified by finding IDs.

GetFindingsStatistics

Lists Amazon GuardDuty findings statistics for the specified detector ID.

GetIPSet

Retrieves the IPSet specified by the ipSetId.

GetInvitationsCount

Returns the count of all GuardDuty membership invitations that were sent to the current

GetMasterAccount

Provides the details for the GuardDuty administrator account associated with the current

GetMemberDetectors

Describes which data sources are enabled for the member account's detector.

GetMembers

Retrieves GuardDuty member accounts (of the current GuardDuty administrator account) specified by

GetThreatIntelSet

Retrieves the ThreatIntelSet that is specified by the ThreatIntelSet ID.

GetUsageStatistics

Lists Amazon GuardDuty usage statistics over the last 30 days for the specified detector

InviteMembers

Invites other AWS accounts (created as members of the current AWS account by

ListDetectors

Lists detectorIds of all the existing Amazon GuardDuty detector resources.

ListFilters

Returns a paginated list of the current filters.

ListFindings

Lists Amazon GuardDuty findings for the specified detector ID.

ListIPSets

Lists the IPSets of the GuardDuty service specified by the detector ID. If you use this

ListInvitations

Lists all GuardDuty membership invitations that were sent to the current AWS

ListMembers

Lists details about all member accounts for the current GuardDuty administrator account.

ListOrganizationAdminAccounts

Lists the accounts configured as GuardDuty delegated administrators.

ListPublishingDestinations

Returns a list of publishing destinations associated with the specified

ListTagsForResource

Lists tags for a resource. Tagging is currently supported for detectors, finding filters,

ListThreatIntelSets

Lists the ThreatIntelSets of the GuardDuty service specified by the detector ID. If you

StartMonitoringMembers

Turns on GuardDuty monitoring of the specified member accounts. Use this operation to

StopMonitoringMembers

Stops GuardDuty monitoring for the specified member accounts. Use the

TagResource

Adds tags to a resource.

UnarchiveFindings

Unarchives GuardDuty findings specified by the findingIds.

UntagResource

Removes tags from a resource.

UpdateDetector

Updates the Amazon GuardDuty detector specified by the detectorId.

UpdateFilter

Updates the filter specified by the filter name.

UpdateFindingsFeedback

Marks the specified GuardDuty findings as useful or not useful.

UpdateIPSet

Updates the IPSet specified by the IPSet ID.

UpdateMemberDetectors

Contains information on member accounts to be updated.

UpdateOrganizationConfiguration

Updates the delegated administrator account with the values provided.

UpdatePublishingDestination

Updates information about the publishing destination specified by the

UpdateThreatIntelSet

Updates the ThreatIntelSet specified by the ThreatIntelSet ID.